+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This changelog follows the [Keep a Changelog](https://keepachangelog.com/en/1.1.
 
 ### Changed
 
+- The output options of all commands are now `--out`. Previously, they varied between `--out`, `--out-dir`, `--output`.
+
 - If `--source` is omitted, it now defaults to the root directory containing the program under coverage. For example, running `cover++ run D:\foo\bar\baz.exe` will use `D:\foo` as the source root.
 
 - Report files now record the Cover++ version used to create them.